Seperated gpio threads
